Add back DEMAND.csv.gz file that was accidentally deleted previously
I am unable to add the lowercase demand.csv.gz file because my OS won't let me save two files with the same spelling (even if one is uppercase and one is lowercase). However, there seems to be a file in the same folder called `demand_MESSAGE.csv.gz` and I wonder if that is the same as `demand.csv.gz`.